T64 codec: fix getMaxCompressedDataSize and test

+++ dbms/tests/queries/0_stateless/00870_t64_codec.sql
@@ -2,12 +2,12 @@ DROP TABLE IF EXISTS test.t64;
 
 CREATE TABLE test.t64
 (
-    u8 UInt64,
-    t_u8 UInt64 Codec(T64, ZSTD),
-    u16 UInt64,
-    t_u16 UInt64 Codec(T64, ZSTD),
-    u32 UInt64,
-    t_u32 UInt64 Codec(T64, ZSTD),
+    u8 UInt8,
+    t_u8 UInt8 Codec(T64, ZSTD),
+    u16 UInt16,
+    t_u16 UInt16 Codec(T64, ZSTD),
+    u32 UInt32,
+    t_u32 UInt32 Codec(T64, ZSTD),
     u64 UInt64,
     t_u64 UInt64 Codec(T64, ZSTD)
 ) ENGINE MergeTree() ORDER BY tuple();
